About to process 250m world data
Lessons learned:
- just go with a square filter for now. Using a bigger one doesn't make Siberia
  any better.
- Don't rescale based on latitude like in the FFT technique, it doesn't make
  things any better either.

+++ textureShading.py
@@ -4,7 +4,7 @@ import numpy.fft as fft
 import scipy.fftpack as scifft
 from osgeo import osr
 
-nextpow2 = lambda v: 2 ** np.ceil(np.log2(v)).astype(int)
+nextpow2 = lambda v: 2 ** int(np.ceil(np.log2(v)))
 
 
 def sciMoreReal(x, alpha, verbose = True):
